+++ oc-includes/osclass/classes/controller/CWebItem.php
@@ -974,7 +974,9 @@ class CWebItem extends BaseModel
     public function doView($file)
     {
         osc_run_hook('before_html');
-        osc_current_web_theme_path($file);
+        if (!osc_gui_page_view($file)) {
+            osc_current_web_theme_path($file);
+        }
         Session::newInstance()->_clearVariables();
         osc_run_hook('after_html');
     }